feat: add evaluation docs and reorganize eval code
Move deepeval/ragas experiments and shared dataset/agent to
python/examples/evaluation/, keeping only adapters in
python/evaluation/. Add evaluation core concepts page to docs.
Remove deprecated python/eval/ directory.
