--- python/eval/agents/requirement/__init__.py
@@ -1,3 +1,0 @@
-# Copyright 2025 © BeeAI a Series of LF Projects, LLC
-# SPDX-License-Identifier: Apache-2.0
-
