--- python/eval/agents/requirement/evaluation_dataset_2_clean.json
@@ -1,28 +1,0 @@
-[
-    {
-      "question": "Which magazine was started first Arthur's Magazine or First for Women?",
-      "answer": "Arthur's Magazine",
-      "relevant_sentences": [
-        "Arthur's Magazine (1844–1846) was an American literary periodical published in Philadelphia in the 19th century.",
-        "First for Women is a woman's magazine published by Bauer Media Group in the USA."
-      ],
-      "wiki_times": 2,
-      "supporting_titles": [
-        "Arthur's Magazine",
-        "First for Women"
-      ]
-    },
-    {
-      "question": "The Oberoi family is part of a hotel company that has a head office in what city?",
-      "answer": "Delhi",
-      "relevant_sentences": [
-        "The Oberoi family is an Indian family that is famous for its involvement in hotels, namely through The Oberoi Group.",
-        "The Oberoi Group is a hotel company with its head office in Delhi."
-      ],
-      "wiki_times": 2,
-      "supporting_titles": [
-        "Oberoi family",
-        "The Oberoi Group"
-      ]
-    }
-  ]
