--- python/eval/setup.py
@@ -1,9 +1,0 @@
-# Copyright 2025 © BeeAI a Series of LF Projects, LLC
-# SPDX-License-Identifier: Apache-2.0
-
-import logging
-
-from dotenv import load_dotenv
-
-logging.disable(logging.DEBUG)
-load_dotenv()
