--- python/evaluation/ragas/data/datasets/my_evaluation.csv
@@ -1,3 +1,0 @@
-id,question,answer,contexts,supporting_titles
-sample_1,Which magazine was started first Arthur's Magazine or First for Women?,Arthur's Magazine,"[""Arthur's Magazine (1844–1846) was an American literary periodical published in Philadelphia in the 19th century."", ""First for Women is a woman's magazine published by Bauer Media Group in the USA.""]","[""Arthur's Magazine"", 'First for Women']"
-sample_2,The Oberoi family is part of a hotel company that has a head office in what city?,Delhi,"['The Oberoi family is an Indian family that is famous for its involvement in hotels, namely through The Oberoi Group.', 'The Oberoi Group is a hotel company with its head office in Delhi.']","['Oberoi family', 'The Oberoi Group']"
