+++ python/examples/evaluation/deepeval/FactsSimilarityMetric.py
@@ -1,16 +1,14 @@
 import sys
 from pathlib import Path
 
-sys.path.insert(0, str(Path(__file__).parent.parent.parent))
+sys.path.insert(0, str(Path(__file__).parent.parent.parent.parent))
 
 from deepeval.metrics import BaseMetric
 from deepeval.test_case import LLMTestCase
 from evaluation.adapters import DeepEvalLLM
 
 
-class AnswerLLMJudgeMetric(BaseMetric):
-    """Uses an LLM as a judge to compare actual vs expected answer semantically."""
-
+class FactsSimilarityMetric(BaseMetric):
     success: bool = False
 
     def __init__(self, model: DeepEvalLLM | None = None, threshold: float = 0.5):
@@ -19,6 +17,12 @@ class AnswerLLMJudgeMetric(BaseMetric):
         self.threshold = threshold
         self.async_mode = True
 
+    def _get_expected(self, test_case: LLMTestCase) -> list[str]:
+        if hasattr(test_case, "expected_facts"):
+            return getattr(test_case, "expected_facts")
+        metadata = getattr(test_case, "additional_metadata", None) or {}
+        return metadata.get("expected_facts", [])
+
     async def a_measure(
         self,
         test_case: LLMTestCase,
@@ -26,32 +30,27 @@ class AnswerLLMJudgeMetric(BaseMetric):
         _in_component: bool = False,
         _log_metric_to_confident: bool = True,
     ) -> float:
-        actual = (test_case.actual_output or "").strip()
-        expected = (test_case.expected_output or "").strip()
-
-        if not expected:
-            score = 1.0 if not actual else 0.0
+        actual_facts = getattr(test_case, "retrieval_context", [])
+        expected_facts = self._get_expected(test_case)
+        if not expected_facts:
+            score = 1.0 if not actual_facts else 0.0
             self.score = score
             self.success = score >= self.threshold
             return score
 
         prompt = (
             "You are an evaluator.\n"
-            "Compare the model's answer to the expected answer.\n\n"
-            f"Question:\n{test_case.input}\n\n"
-            f"Model answer:\n{actual}\n\n"
-            f"Expected answer:\n{expected}\n\n"
+            "Compare the two lists of supporting facts.\n\n"
+            f"Actual facts:\n{actual_facts}\n\n"
+            f"Expected facts:\n{expected_facts}\n\n"
             "Return ONLY a number between 0 and 1 (no text), where:\n"
-            "0 = completely incorrect or unrelated,\n"
-            "1 = fully correct and equivalent in meaning.\n"
+            "0 = completely different, 1 = identical in meaning.\n"
         )
-
         text = await self.model.a_generate(prompt)
         try:
             score = float(str(text).strip())
         except Exception:
             score = 0.0
-
         score = max(0.0, min(1.0, score))
         self.score = score
         self.success = score >= self.threshold
@@ -65,5 +64,5 @@ class AnswerLLMJudgeMetric(BaseMetric):
         return getattr(self, "success", False)
 
     @property
-    def __name__(self) -> str:
-        return "AnswerLLMJudgeMetric"
+    def __name__(self):
+        return "FactsSimilarityMetric"
