+++ python/examples/evaluation/deepeval/ToolUsageMetric.py
@@ -1,7 +1,7 @@
 import sys
 from pathlib import Path
 
-sys.path.insert(0, str(Path(__file__).parent.parent.parent))
+sys.path.insert(0, str(Path(__file__).parent.parent.parent.parent))
 
 from deepeval.metrics import BaseMetric
 from deepeval.test_case import LLMTestCase
